+++ app/id/[slug]/page.tsx
@@ -12,21 +12,23 @@ import { Divider } from '@nextui-org/divider';
 import { Skeleton } from '@nextui-org/skeleton';
 import { useCallback, useState } from 'react';
 
+import { QuestionBody } from '@/components/profile/questions/questionBody';
 import { SkeltonProfile } from '@/components/profile/skeleton';
 import { useProfileData } from '@/hooks/useQuestions';
 import type { QuestionsType } from '@/types/questionType';
+import { Image } from '@nextui-org/image';
+import { Modal, ModalBody, ModalContent } from '@nextui-org/modal'; // Import a modal component from your UI library
 
 export default function ProfilePage({ params }: { params: { slug: string } }) {
 	const currentUsername = params.slug;
 	const { user, isLoading, questions, isError } =
 		useProfileData(currentUsername);
-	
-  
+
 	const [showComments, setShowComments] = useState<Record<string, boolean>>({});
+	const [selectedImage, setSelectedImage] = useState<string | null>(null);
 	const [isShareModalOpen, setIsShareModalOpen] = useState<
 		Record<string, boolean>
 	>({});
-
 
 	const toggleComments = useCallback((questionId: string) => {
 		setShowComments((prev) => ({ ...prev, [questionId]: !prev[questionId] }));
@@ -38,25 +40,29 @@ export default function ProfilePage({ params }: { params: { slug: string } }) {
 		}));
 	}, []);
 
+	const handleImageClick = (imageUrl: string) => {
+		setSelectedImage(imageUrl);
+	};
+
+	const handleCloseModal = () => {
+		setSelectedImage(null);
+	};
+
 	if (isLoading) {
 		<SkeltonProfile />;
 	}
-
+  
 	return (
-    <div className="max-w-4xl mx-auto px-4 sm:px-6 lg:px-8">
+		<div className="max-w-4xl mx-auto px-4 sm:px-6 lg:px-8">
 			{/* Main Card */}
-      <div className="mt-4">
-				<UserProfileBox
-					username={currentUsername}
-          isSingleQuestion={false}
-				/>
+			<div className="mt-4">
+				<UserProfileBox username={currentUsername} isSingleQuestion={false} />
 			</div>
-      <div className="mt-4">
-					<QuestionBox username={currentUsername} user={user} />
-				</div>
-
+			<div className="mt-4">
+				<QuestionBox username={currentUsername} user={user} />
+			</div>
 			{/* Question List Box */}
-      <div className="flex flex-col w-full space-y-4 mt-4">
+			<div className="flex flex-col w-full space-y-4 mt-4">
 				{isError ? (
 					<Card className="flex flex-col items-center justify-center py-4 px-4">
 						<p className="text-secondary text-center text-lg font-semibold mb-2">
@@ -81,7 +87,7 @@ export default function ProfilePage({ params }: { params: { slug: string } }) {
 					questions.map((question: QuestionsType) => (
 						<Card key={question.questionId} className="p-2 sm:p-4">
 							<QuestionHeader
-                isAnon={question.isAnon}  
+								isAnon={question.isAnon}
 								posterId={question.posterId}
 								questionId={question.questionId}
 								posterUsername={question.posterUsername}
@@ -91,9 +97,11 @@ export default function ProfilePage({ params }: { params: { slug: string } }) {
 								user={user}
 							/>
 							<CardBody className="py-1 sm:py-2">
-								<p className="text-sm sm:text-base line-clamp-3">
-									{question.content}
-								</p>
+								<QuestionBody
+									content={question.content}
+									imageUrls={question.imageUrls || []}
+									onImageClick={handleImageClick}
+								/>
 							</CardBody>
 							<CardFooter className="flex flex-col items-start pt-1 sm:pt-2">
 								<QuestionStats
@@ -118,7 +126,7 @@ export default function ProfilePage({ params }: { params: { slug: string } }) {
 										questionId={question.questionId}
 										user={user}
 										isSingleQuestion={false}
-                    username={currentUsername}
+										username={currentUsername}
 									/>
 								)}
 							</CardFooter>
@@ -133,6 +141,29 @@ export default function ProfilePage({ params }: { params: { slug: string } }) {
 					))
 				)}
 			</div>
-    </div>
+			{selectedImage && (
+				<Modal
+					isOpen={!!selectedImage}
+					onClose={handleCloseModal}
+					backdrop="blur"
+					size="xl"
+					className="bg-transparent flex items-center shadow-none"
+					placement="center"
+				>
+					<ModalContent>
+						{(onClose) => (
+							<ModalBody className="p-0 overflow-hidden rounded-lg">
+								<Image
+									src={selectedImage || ''}
+									alt="Zoomed Image"
+									className="max-w-full max-h-[80vh] object-contain"
+									onClick={onClose}
+								/>
+							</ModalBody>
+						)}
+					</ModalContent>
+				</Modal>
+			)}
+		</div>
 	);
 }